% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeights.R
\name{getWeights}
\alias{getWeights}
\title{getWeights}
\usage{
getWeights(counts, selectTissue)
}
\arguments{
\item{counts}{Set counts equal to a file path of an document (such as excel file) containing a normalized count matrix with markers on the y axis and organ/tissues as the header row. Before running the function, take a look at the provided packageWeights.xlsx file and note the organs/tissues listed here. Going back to your counts file, create a second row underneath your header row. in this row, for each element in the header row, type in an equivalent tissue from the list of tissues in the packageWeights file. For example, if cell B1 was "heart left ventricle" the closest approximation in the packageWeights file is "heart", so heart would go in cell B2. And so on. Import this modified counts file as the parameter for the function.}

\item{selectTissue}{choose a tissue from the list of tissues included in the packageWeights.xlsx file. Import as a string.}
}
\description{
This function generates a list of weights that can be used in the markerScore function.
}
\keyword{specificity}
