% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markerScore.R
\name{markerScore}
\alias{markerScore}
\title{markerScore}
\usage{
markerScore(counts, gene, weightsforTissue)
}
\arguments{
\item{counts}{Set counts equal to a file path of an document (such as excel file) containing a normalized count matrix with markers (ENSEMBL ID works best) on the y axis and organ/tissues as the header row.}

\item{gene}{Set gene equal to an id for a transcript marker (such as ENSEMBL ID)}

\item{weightsforTissue}{Set this equal to the output of the getWeights function.}
}
\description{
This function gives a probability for how specific a marker is for a organ/tissue.
}
